Optimize code by removing unused variables and arguments
Several tests within the 'opendsu-sdk' had variables declared that were not being utilized, and callback functions included arguments that were not needed. The unnecessary variables and arguments have been removed to simplify and streamline the code. This should not affect any functionality and is primarily a matter of developers' preference for clean, efficient code.